Gets webpack working with jsx files
