+++ app/client/components/Canvas.jsx
